chore: clean fee utils (#1968)
