+++ crates/blockifier/src/fee/actual_cost.rs
@@ -82,9 +82,14 @@ impl TransactionReceipt {
             n_reverted_steps: reverted_steps,
         };
 
+        let gas = tx_resources.to_gas_vector(
+            &tx_context.block_context.versioned_constants,
+            tx_context.block_context.block_info.use_kzg_da,
+        )?;
+
         // L1 handler transactions are not charged an L2 fee but it is compared to the L1 fee.
         let fee = if tx_context.tx_info.enforce_fee()? || tx_type == TransactionType::L1Handler {
-            tx_context.tx_info.calculate_tx_fee(&tx_resources, &tx_context.block_context)?
+            tx_context.tx_info.get_fee_by_gas_vector(&tx_context.block_context.block_info, gas)
         } else {
             Fee(0)
         };
@@ -92,10 +97,6 @@ impl TransactionReceipt {
             .starknet_resources
             .get_state_changes_cost(tx_context.block_context.block_info.use_kzg_da);
 
-        let gas = tx_resources.to_gas_vector(
-            &tx_context.block_context.versioned_constants,
-            tx_context.block_context.block_info.use_kzg_da,
-        )?;
         Ok(Self { resources: tx_resources, gas, da_gas, fee })
     }
